Re-aligning Shed and MineEntrance
- Changed horizontal anchoring for Shed and MineEntrance from sides of the screen to the middle of the screen.

These 2 objects were horizontally anchored to the sides of the screen which meant different distance between them on different resolutions. And it's unclear whether we want to stretch the mineshafts underground with wider screens or not, they're not even there yet.

So for now we don't want to deal with any of this, these objects were re-anchored to the middle of the screen such that no matter the resolution, they'll be the same distance apart.

We're gonna keep the scene simple for now and the mineshaft under those objects won't have to be horizontally stretched with the new anchors.